Package naming refactor
