+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="shop.fast_shop"
     xmlns:tools="http://schemas.android.com/tools">
 
     <application
@@ -13,7 +14,7 @@
         android:theme="@style/Theme.Listadecompras"
         tools:targetApi="31">
         <activity
-            android:name="rrff.fast_shop.MainActivity"
+            android:name="shop.fast_shop.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Listadecompras">
