--- app/src/main/java/rrff/fast_shop/model/LoginState.kt
@@ -1,8 +1,0 @@
-package rrff.fast_shop.model
-
-data class LoginState(
-    val email: String = "",
-    val password: String = "",
-    val passwordVisible: Boolean = false,
-    val showError: Boolean = false
-)
