--- app/src/main/java/rrff/fast_shop/navigation/Routes.kt
@@ -1,8 +1,0 @@
-package rrff.fast_shop.navigation
-
-sealed class Screen(val route: String) {
-    object Login : Screen("login")
-    object SignUp : Screen("SignUp")
-    object Home : Screen("HomeScreen")
-
-}
