--- app/src/main/java/rrff/fast_shop/service/AuthService.kt
@@ -1,43 +1,0 @@
-package rrff.fast_shop.service // Atualize o pacote
-
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
-
-class AuthService {
-    private val auth: FirebaseAuth = Firebase.auth
-
-    // Login com email/senha
-    fun login(
-        email: String,
-        password: String,
-        onSuccess: () -> Unit,
-        onFailure: (String) -> Unit
-    ) {
-        auth.signInWithEmailAndPassword(email, password)
-            .addOnCompleteListener { task ->
-                if (task.isSuccessful) {
-                    onSuccess()
-                } else {
-                    onFailure(task.exception?.message ?: "Erro desconhecido")
-                }
-            }
-    }
-
-    // Cadastro de novo usuário
-    fun register(
-        email: String,
-        password: String,
-        onSuccess: () -> Unit,
-        onFailure: (String) -> Unit
-    ) {
-        auth.createUserWithEmailAndPassword(email, password)
-            .addOnCompleteListener { task ->
-                if (task.isSuccessful) {
-                    onSuccess()
-                } else {
-                    onFailure(task.exception?.message ?: "Erro ao criar conta")
-                }
-            }
-    }
-}
