--- app/src/main/java/rrff/fast_shop/ui/theme/Color.kt
@@ -1,331 +1,0 @@
-package rrff.fast_shop.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-
-/* Red */
-val Red = Color( 0xfff44336)
-val Red50 = Color( 0xffffebee)
-val Red100 = Color( 0xffffcdd2)
-val Red200 = Color( 0xffef9a9a)
-val Red300 = Color( 0xffe57373)
-val Red400 = Color( 0xffef5350)
-val Red500 = Color( 0xfff44336)
-val Red600 = Color( 0xffe53935)
-val Red700 = Color( 0xffd32f2f)
-val Red800 = Color( 0xffc62828)
-val Red900 = Color( 0xffb71c1c)
-val RedA100 = Color( 0xffff8a80)
-val RedA200 = Color( 0xffff5252)
-val RedA400 = Color( 0xffff1744)
-val RedA700 = Color( 0xffd50000)
-
-/* Pink */
-val Pink = Color( 0xffe91e63)
-val Pink50 = Color( 0xfffce4ec)
-val Pink100 = Color( 0xfff8bbd0)
-val Pink200 = Color( 0xfff48fb1)
-val Pink300 = Color( 0xfff06292)
-val Pink400 = Color( 0xffec407a)
-val Pink500 = Color( 0xffe91e63)
-val Pink600 = Color( 0xffd81b60)
-val Pink700 = Color( 0xffc2185b)
-val Pink800 = Color( 0xffad1457)
-val Pink900 = Color( 0xff880e4f)
-val PinkA100 = Color( 0xffff80ab)
-val PinkA200 = Color( 0xffff4081)
-val PinkA400 = Color( 0xfff50057)
-val PinkA700 = Color( 0xffc51162)
-
-/* Purple */
-val Purple = Color( 0xff9c27b0)
-val Purple50 = Color( 0xfff3e5f5)
-val Purple100 = Color( 0xffe1bee7)
-val Purple200 = Color( 0xffce93d8)
-val Purple300 = Color( 0xffba68c8)
-val Purple400 = Color( 0xffab47bc)
-val Purple500 = Color( 0xff9c27b0)
-val Purple600 = Color( 0xff8e24aa)
-val Purple700 = Color( 0xff7b1fa2)
-val Purple800 = Color( 0xff6a1b9a)
-val Purple900 = Color( 0xff4a148c)
-val PurpleA100 = Color( 0xffea80fc)
-val PurpleA200 = Color( 0xffe040fb)
-val PurpleA400 = Color( 0xffd500f9)
-val PurpleA700 = Color( 0xffaa00ff)
-
-/* Deep Purple */
-val DeepPurple = Color( 0xff673ab7)
-val DeepPurple50 = Color( 0xffede7f6)
-val DeepPurple100 = Color( 0xffd1c4e9)
-val DeepPurple200 = Color( 0xffb39ddb)
-val DeepPurple300 = Color( 0xff9575cd)
-val DeepPurple400 = Color( 0xff7e57c2)
-val DeepPurple500 = Color( 0xff673ab7)
-val DeepPurple600 = Color( 0xff5e35b1)
-val DeepPurple700 = Color( 0xff512da8)
-val DeepPurple800 = Color( 0xff4527a0)
-val DeepPurple900 = Color( 0xff311b92)
-val DeepPurpleA100 = Color( 0xffb388ff)
-val DeepPurpleA200 = Color( 0xff7c4dff)
-val DeepPurpleA400 = Color( 0xff651fff)
-val DeepPurpleA700 = Color( 0xff6200ea)
-
-/* Indigo */
-val Indigo = Color( 0xff3f51b5)
-val Indigo50 = Color( 0xffe8eaf6)
-val Indigo100 = Color( 0xffc5cae9)
-val Indigo200 = Color( 0xff9fa8da)
-val Indigo300 = Color( 0xff7986cb)
-val Indigo400 = Color( 0xff5c6bc0)
-val Indigo500 = Color( 0xff3f51b5)
-val Indigo600 = Color( 0xff3949ab)
-val Indigo700 = Color( 0xff303f9f)
-val Indigo800 = Color( 0xff283593)
-val Indigo900 = Color( 0xff1a237e)
-val IndigoA100 = Color( 0xff8c9eff)
-val IndigoA200 = Color( 0xff536dfe)
-val IndigoA400 = Color( 0xff3d5afe)
-val IndigoA700 = Color( 0xff304ffe)
-
-/* Blue */
-val Blue = Color( 0xff2196f3)
-val Blue50 = Color( 0xffe3f2fd)
-val Blue100 = Color( 0xffbbdefb)
-val Blue200 = Color( 0xff90caf9)
-val Blue300 = Color( 0xff64b5f6)
-val Blue400 = Color( 0xff42a5f5)
-val Blue500 = Color( 0xff2196f3)
-val Blue600 = Color( 0xff1e88e5)
-val Blue700 = Color( 0xff1976d2)
-val Blue800 = Color( 0xff1565c0)
-val Blue900 = Color( 0xff0d47a1)
-val BlueA100 = Color( 0xff82b1ff)
-val BlueA200 = Color( 0xff448aff)
-val BlueA400 = Color( 0xff2979ff)
-val BlueA700 = Color( 0xff2962ff)
-
-/* Light Blue */
-val LightBlue = Color( 0xff03a9f4)
-val LightBlue50 = Color( 0xffe1f5fe)
-val LightBlue100 = Color( 0xffb3e5fc)
-val LightBlue200 = Color( 0xff81d4fa)
-val LightBlue300 = Color( 0xff4fc3f7)
-val LightBlue400 = Color( 0xff29b6f6)
-val LightBlue500 = Color( 0xff03a9f4)
-val LightBlue600 = Color( 0xff039be5)
-val LightBlue700 = Color( 0xff0288d1)
-val LightBlue800 = Color( 0xff0277bd)
-val LightBlue900 = Color( 0xff01579b)
-val LightBlueA100 = Color( 0xff80d8ff)
-val LightBlueA200 = Color( 0xff40c4ff)
-val LightBlueA400 = Color( 0xff00b0ff)
-val LightBlueA700 = Color( 0xff0091ea)
-
-/* Cyan */
-val Cyan = Color( 0xff00bcd4)
-val Cyan50 = Color( 0xffe0f7fa)
-val Cyan100 = Color( 0xffb2ebf2)
-val Cyan200 = Color( 0xff80deea)
-val Cyan300 = Color( 0xff4dd0e1)
-val Cyan400 = Color( 0xff26c6da)
-val Cyan500 = Color( 0xff00bcd4)
-val Cyan600 = Color( 0xff00acc1)
-val Cyan700 = Color( 0xff0097a7)
-val Cyan800 = Color( 0xff00838f)
-val Cyan900 = Color( 0xff006064)
-val CyanA100 = Color( 0xff84ffff)
-val CyanA200 = Color( 0xff18ffff)
-val CyanA400 = Color( 0xff00e5ff)
-val CyanA700 = Color( 0xff00b8d4)
-
-/* Teal */
-val Teal = Color( 0xff009688)
-val Teal50 = Color( 0xffe0f2f1)
-val Teal100 = Color( 0xffb2dfdb)
-val Teal200 = Color( 0xff80cbc4)
-val Teal300 = Color( 0xff4db6ac)
-val Teal400 = Color( 0xff26a69a)
-val Teal500 = Color( 0xff009688)
-val Teal600 = Color( 0xff00897b)
-val Teal700 = Color( 0xff00796b)
-val Teal800 = Color( 0xff00695c)
-val Teal900 = Color( 0xff004d40)
-val TealA100 = Color( 0xffa7ffeb)
-val TealA200 = Color( 0xff64ffda)
-val TealA400 = Color( 0xff1de9b6)
-val TealA700 = Color( 0xff00bfa5)
-
-/* Green */
-val Green = Color( 0xff4caf50)
-val Green50 = Color( 0xffe8f5e9)
-val Green100 = Color( 0xffc8e6c9)
-val Green200 = Color( 0xffa5d6a7)
-val Green300 = Color( 0xff81c784)
-val Green400 = Color( 0xff66bb6a)
-val Green500 = Color( 0xff4caf50)
-val Green600 = Color( 0xff43a047)
-val Green700 = Color( 0xff388e3c)
-val Green800 = Color( 0xff2e7d32)
-val Green900 = Color( 0xff1b5e20)
-val GreenA100 = Color( 0xffb9f6ca)
-val GreenA200 = Color( 0xff69f0ae)
-val GreenA400 = Color( 0xff00e676)
-val GreenA700 = Color( 0xff00c853)
-
-/* Light Green */
-val LightGreen = Color( 0xff8bc34a)
-val LightGreen50 = Color( 0xfff1f8e9)
-val LightGreen100 = Color( 0xffdcedc8)
-val LightGreen200 = Color( 0xffc5e1a5)
-val LightGreen300 = Color( 0xffaed581)
-val LightGreen400 = Color( 0xff9ccc65)
-val LightGreen500 = Color( 0xff8bc34a)
-val LightGreen600 = Color( 0xff7cb342)
-val LightGreen700 = Color( 0xff689f38)
-val LightGreen800 = Color( 0xff558b2f)
-val LightGreen900 = Color( 0xff33691e)
-val LightGreenA100 = Color( 0xffccff90)
-val LightGreenA200 = Color( 0xffb2ff59)
-val LightGreenA400 = Color( 0xff76ff03)
-val LightGreenA700 = Color( 0xff64dd17)
-
-/* Lime */
-val Lime = Color( 0xffcddc39)
-val Lime50 = Color( 0xfff9fbe7)
-val Lime100 = Color( 0xfff0f4c3)
-val Lime200 = Color( 0xffe6ee9c)
-val Lime300 = Color( 0xffdce775)
-val Lime400 = Color( 0xffd4e157)
-val Lime500 = Color( 0xffcddc39)
-val Lime600 = Color( 0xffc0ca33)
-val Lime700 = Color( 0xffafb42b)
-val Lime800 = Color( 0xff9e9d24)
-val Lime900 = Color( 0xff827717)
-val LimeA100 = Color( 0xfff4ff81)
-val LimeA200 = Color( 0xffeeff41)
-val LimeA400 = Color( 0xffc6ff00)
-val LimeA700 = Color( 0xffaeea00)
-
-/* Yellow */
-val Yellow = Color( 0xffffeb3b)
-val Yellow50 = Color( 0xfffffde7)
-val Yellow100 = Color( 0xfffff9c4)
-val Yellow200 = Color( 0xfffff59d)
-val Yellow300 = Color( 0xfffff176)
-val Yellow400 = Color( 0xffffee58)
-val Yellow500 = Color( 0xffffeb3b)
-val Yellow600 = Color( 0xfffdd835)
-val Yellow700 = Color( 0xfffbc02d)
-val Yellow800 = Color( 0xfff9a825)
-val Yellow900 = Color( 0xfff57f17)
-val YellowA100 = Color( 0xffffff8d)
-val YellowA200 = Color( 0xffffff00)
-val YellowA400 = Color( 0xffffea00)
-val YellowA700 = Color( 0xffffd600)
-
-/* Amber */
-val Amber = Color( 0xffffc107)
-val Amber50 = Color( 0xfffff8e1)
-val Amber100 = Color( 0xffffecb3)
-val Amber200 = Color( 0xffffe082)
-val Amber300 = Color( 0xffffd54f)
-val Amber400 = Color( 0xffffca28)
-val Amber500 = Color( 0xffffc107)
-val Amber600 = Color( 0xffffb300)
-val Amber700 = Color( 0xffffa000)
-val Amber800 = Color( 0xffff8f00)
-val Amber900 = Color( 0xffff6f00)
-val AmberA100 = Color( 0xffffe57f)
-val AmberA200 = Color( 0xffffd740)
-val AmberA400 = Color( 0xffffc400)
-val AmberA700 = Color( 0xffffab00)
-
-/* Orange */
-val Orange = Color( 0xffff9800)
-val Orange50 = Color( 0xfffff3e0)
-val Orange100 = Color( 0xffffe0b2)
-val Orange200 = Color( 0xffffcc80)
-val Orange300 = Color( 0xffffb74d)
-val Orange400 = Color( 0xffffa726)
-val Orange500 = Color( 0xffff9800)
-val Orange600 = Color( 0xfffb8c00)
-val Orange700 = Color( 0xfff57c00)
-val Orange800 = Color( 0xffef6c00)
-val Orange900 = Color( 0xffe65100)
-val OrangeA100 = Color( 0xffffd180)
-val OrangeA200 = Color( 0xffffab40)
-val OrangeA400 = Color( 0xffff9100)
-val OrangeA700 = Color( 0xffff6d00)
-
-/* Deep Orange */
-val DeepOrange = Color( 0xffff5722)
-val DeepOrange50 = Color( 0xfffbe9e7)
-val DeepOrange100 = Color( 0xffffccbc)
-val DeepOrange200 = Color( 0xffffab91)
-val DeepOrange300 = Color( 0xffff8a65)
-val DeepOrange400 = Color( 0xffff7043)
-val DeepOrange500 = Color( 0xffff5722)
-val DeepOrange600 = Color( 0xfff4511e)
-val DeepOrange700 = Color( 0xffe64a19)
-val DeepOrange800 = Color( 0xffd84315)
-val DeepOrange900 = Color( 0xffbf360c)
-val DeepOrangeA100 = Color( 0xffff9e80)
-val DeepOrangeA200 = Color( 0xffff6e40)
-val DeepOrangeA400 = Color( 0xffff3d00)
-val DeepOrangeA700 = Color( 0xffdd2c00)
-
-/* Brown */
-val Brown = Color( 0xff795548)
-val Brown50 = Color( 0xffefebe9)
-val Brown100 = Color( 0xffd7ccc8)
-val Brown200 = Color( 0xffbcaaa4)
-val Brown300 = Color( 0xffa1887f)
-val Brown400 = Color( 0xff8d6e63)
-val Brown500 = Color( 0xff795548)
-val Brown600 = Color( 0xff6d4c41)
-val Brown700 = Color( 0xff5d4037)
-val Brown800 = Color( 0xff4e342e)
-val Brown900 = Color( 0xff3e2723)
-
-/* Grey */
-val Grey = Color( 0xff9e9e9e)
-val Grey50 = Color( 0xfffafafa)
-val Grey100 = Color( 0xfff5f5f5)
-val Grey200 = Color( 0xffeeeeee)
-val Grey300 = Color( 0xffe0e0e0)
-val Grey400 = Color( 0xffbdbdbd)
-val Grey500 = Color( 0xff9e9e9e)
-val Grey600 = Color( 0xff757575)
-val Grey700 = Color( 0xff616161)
-val Grey800 = Color( 0xff424242)
-val Grey900 = Color( 0xff212121)
-
-/* Blue Grey */
-val BlueGrey = Color( 0xff607d8b)
-val BlueGrey50 = Color( 0xffeceff1)
-val BlueGrey100 = Color( 0xffcfd8dc)
-val BlueGrey200 = Color( 0xffb0bec5)
-val BlueGrey300 = Color( 0xff90a4ae)
-val BlueGrey400 = Color( 0xff78909c)
-val BlueGrey500 = Color( 0xff607d8b)
-val BlueGrey600 = Color( 0xff546e7a)
-val BlueGrey700 = Color( 0xff455a64)
-val BlueGrey800 = Color( 0xff37474f)
-val BlueGrey900 = Color( 0xff263238)
-
-/* White / Black */
-val white = Color( 0xffffffff)
-val black = Color( 0xff000000)
-
-/* Monk skin tone */
-val monk01 = Color( 0xfff6ede4)
-val monk02 = Color( 0xfff3e7db)
-val monk03 = Color( 0xfff7ead0)
-val monk04 = Color( 0xffeadaba)
-val monk05 = Color( 0xffd7bd96)
-val monk06 = Color( 0xffa07e56)
-val monk07 = Color( 0xff825c43)
-val monk08 = Color( 0xff604134)
-val monk09 = Color( 0xff3a312a)
-val monk10 = Color( 0xff292420)
