--- app/src/main/java/rrff/fast_shop/ui/theme/Typography.kt
@@ -1,10 +1,0 @@
-package rrff.fast_shop.ui.theme
-
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
-import rrff.fast_shop.R.font
-
-val WorkSansFontFamily = FontFamily(
-    Font(font.work_sans),
-
-)
