--- app/src/main/java/rrff/fast_shop/viewmodel/HomeViewModel.kt
@@ -1,1 +1,0 @@
-package rrff.fast_shop.viewmodel
