+++ app/src/main/java/shop/fast_shop/model/LoginState.kt
@@ -1,0 +1,8 @@
+package shop.fast_shop.model
+
+data class LoginState(
+    val email: String = "",
+    val password: String = "",
+    val passwordVisible: Boolean = false,
+    val showError: Boolean = false
+)
