+++ app/src/main/java/shop/fast_shop/model/SignUpState.kt
@@ -1,0 +1,10 @@
+package shop.fast_shop.model
+
+data class SignUpState(
+    val email: String = "",
+    val password: String = "",
+    val confirmPassword: String = "",
+    val passwordVisible: Boolean = false,
+    val showError: Boolean = false,
+    val errorMessage: String = ""
+)
