+++ app/src/main/java/shop/fast_shop/navigation/Routes.kt
@@ -1,0 +1,8 @@
+package shop.fast_shop.navigation
+
+sealed class Screen(val route: String) {
+    object Login : Screen("login")
+    object SignUp : Screen("SignUp")
+    object Home : Screen("HomeScreen")
+
+}
