+++ app/src/main/java/shop/fast_shop/ui/theme/Typography.kt
@@ -1,0 +1,5 @@
+package shop.fast_shop.ui.theme
+
+import androidx.compose.ui.text.font.Font
+import androidx.compose.ui.text.font.FontFamily
+
