+++ app/src/main/java/shop/fast_shop/viewmodel/HomeViewModel.kt
@@ -1,0 +1,1 @@
+package shop.fast_shop.viewmodel
